Crianda a interacao entre back e front

+++ backend/backend/src/main/java/com/bmss/backend/repository/RoleRepository.java
@@ -3,5 +3,4 @@ package com.bmss.backend.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.bmss.backend.model.Role;
 
-public interface RoleRepository extends JpaRepository<Role, Long> {
-}
+public interface RoleRepository extends JpaRepository<Role, Integer> {}
