+++ backend/backend/src/main/java/com/bmss/backend/repository/SentimentRepository.java
@@ -3,5 +3,4 @@ package com.bmss.backend.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.bmss.backend.model.Sentiment;
 
-public interface SentimentRepository extends JpaRepository<Sentiment, Long> {
-}
+public interface SentimentRepository extends JpaRepository<Sentiment, Integer> {}
