+++ backend/backend/src/main/java/com/bmss/backend/repository/SourceRepository.java
@@ -3,5 +3,4 @@ package com.bmss.backend.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.bmss.backend.model.Source;
 
-public interface SourceRepository extends JpaRepository<Source, Long> {
-}
+public interface SourceRepository extends JpaRepository<Source, Integer> {}
